ts: address review comments
